Refactor: 테스트 파일 구조 개선
